+++ sources/core/objects/PhysicsEngine.h
@@ -27,8 +27,8 @@ public:
 
     // TODO: base force vector value for movement on 1 tile should more than 1.
     // So force vectors values should be scaled up
-    static Dir ProcessForceTick(Vector* force, int* progress, int friction, int mass);
-    static void ApplyForce(Vector* force, int* progress, const Vector& addition);
+    static Dir ProcessForceTick(Vector* force, qint64* progress, int friction, int mass);
+    static void ApplyForce(Vector* force, qint64* progress, const Vector& addition);
 private:
     void Clear();
